switch from enums to structs

+++ Sources/Mach/Core/SpecialPorts/Bootstrap/ConnectionPort.swift
@@ -24,7 +24,7 @@ extension Mach {
             var options = mach_port_options_t()
             options.service_port_name = servicePort.name
             options.mpl = limits
-            options.flags = flags.bitmap()
+            options.flags = UInt32(flags.bitmap())
             try self.init(options: options, context: context, inNameSpaceOf: task)
         }
